requirements.txt
requests
rich
python-dotenv
beautifulsoup4
